Try to find EXIF orientation from all files, not only JPEGs
EXIF support was added to the PNG spec in July 2017, and tools like
ImageMagick (7+) and Exiftool now have it in place. Thus there is a
chance that orientation info can be found even if the file is not a
JPEG and there's no JPEG by the same name in the current directory.