Convert uint16 textures to float16 instead of float32
Conversion to float is necessary due to a ModernGL bug/limitation
(as of version 5.5.2) that causes integer textures to yield only
zeros when sampled in the fragment shader.

Previously, we were converting to float32 to preserve full 16-bit
precision, but float16 is a better tradeoff with ~11-bit precision
at half the size.