[imageprovider] No need to sleep between loading each image
The loader is already running in a separate thread, so there is no
need for it to yield. The hardcoded 10ms sleep was capping maximum
file loading rate at 100 per second, which becomes a bottleneck
when there are hundreds or thousands of (very small) files.

Also, filenames are no longer printed out if there are more than
200 files, unless the '--verbose' flag is given. This is to avoid
cluttering the console in regular use.

This commit constitutes release 1.2.1.