[glrenderer] Speed up _fit_gamut() by aggressive downscaling
Pick only every Nth pixel in each direction, where N depends on
the number of megapixels on a logarithmic scale. For example, a
100MP image is downsampled by a factor of 8x8.

This does run the risk of missing very small out-of-gamut details,
but on the other hand, processing at full resolution makes loading
of large images annoyingly slow.

Also switch from float64 to float32 for further speed-up. Extreme
numeric range is not really needed.