[pygletui] Switch to manual keyboard state tracking
On Windows, the operating system is stealing arrow key events
while the Windows key is pressed. This causes the image to move
around on its own, which is highly annoying. The workaround is
to force the arrow keys into "released" state while the Windows
key remains down.

This commit constitutes release 1.9.1.